chore: move types, small refactor
